feat(core/rust): introduce Wordlist component for both bip39 and slip39
